initial index and App strucutre rework
